requirements.txt
fastapi
uvicorn
python-dotenv
langchain==0.2.16
langchain-community
langchain-huggingface
langchain-google-genai
pinecone-client
langchain-pinecone
sentence-transformers
pytest
httpx
